pyproject.toml
[tool.poetry]
name = "podcast-cuts"
version = "0.1.0"
description = "Sistema automatizado para criar e publicar cortes de podcasts"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
yt-dlp = "^2025.5.22"
faster-whisper = "^1.1.1"
moviepy = "^1.0.3"
python-dotenv = "^1.1.0"
PyYAML = "^6.0.2"
google-api-python-client = "^2.170.0"
google-auth-oauthlib = "^1.2.2"
pillow = "^11.2.1"
imageio-ffmpeg = "^0.6.0"
opencv-python = "^4.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 